Add async func one

+++ 0x01-python_async_function/0-basic_async_syntax.py
@@ -1,12 +1,10 @@
 #!/usr/bin/env python3
 """Asynch I/O"""
-
-
 import random
 import asyncio
 
 
-async def wait_random(max_delay: float=10.0) -> float:
+async def wait_random(max_delay: int=10) -> float:
     """
     Generate a random float 
     implement async funcs
